move the calci.js file to src folder.
